+++ ahead/ARMAGARCH/ArmaGarch.py
@@ -101,26 +101,6 @@ class ArmaGarch(object):
         sims_: a numpy array
             forecasting simulations
 
-    Examples:
-
-        ```
-            import pandas as pd
-            from ahead import ArmaGarch
-
-            # Data frame containing the time series
-            dataset = {
-            'date' : ['2020-01-01', '2020-02-01', '2020-03-01', '2020-04-01', '2020-05-01'],
-            'value' : [34, 30, 35.6, 33.3, 38.1]}
-
-            df = pd.DataFrame(dataset).set_index('date')
-            print(df)
-
-            # univariate time series forecasting
-            e1 = ArmaGarch(h = 5) 
-            e1.forecast(df)
-            print(e1.result_df_)
-        ```
-
     """
 
     def __init__(