+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 - add dropout regularization to `Ridge2Regressor`
 - parallel execution for type_pi == bootstrap in `Ridge2Regressor` (done in R /!\, experimental)
+- ARMA(1, 1)-GARCH(1, 1) in Python 
 
 # version 0.4.2